==> 2.CalculateOffsets.Rmd <==
---
title: "2.CalcualteOffsets"
output: html_document
date: "2024-04-02"
---

```{r, setup, include=FALSE}
knitr::opts_knit$set(root.dir=".")
```


This document is based on the workflow produce by Elly Knight to produce BAM's National models v5 and it is basically an exact copy of the version produce for the National models v5 (02.CaculateOffsets) with the modifications of the pacakge wildRtrax using directly the code to create a new function qpad_offsets to extract the offsets to be able to integrate ebird data (as wt_qpad_offsets, only works for wildtrax data). 

Be aware that you need to have QPAD repository in your local.

#0. Prep

```{r}
library(tidyverse) #basic data wrangling
library(lubridate) #temporal data wrangling
library(QPAD) #to get offset model estimates
library(terra)
library(intrval) #required for QPAD
library(data.table) #collapse list to dataframe
library(reproducible)
```

Set root path 
```{r}
root <- "."
```


#1. Load QPAD requirements

##1.1 Estimates
```{r}
load_BAM_QPAD(version = 3)
```

##1.2 Raster data

Get tifs from repo.
```{r}

rlcc<- prepInputs(url = "https://raw.githubusercontent.com/ABbiodiversity/wildRtrax-assets/main/lcc.tif", 
                  destinationPath = file.path(root, "Offset_prep"))
rtree <- prepInputs(url ="https://raw.githubusercontent.com/ABbiodiversity/wildRtrax-assets/main/tree.tif", 
                    destinationPath = file.path(root, "Offset_prep"))
rd1 <-  prepInputs(url ="https://raw.githubusercontent.com/ABbiodiversity/wildRtrax-assets/main/seedgrow.tif",
                   destinationPath = file.path(root, "Offset_prep"))
rtz <-  prepInputs (url ="https://raw.githubusercontent.com/ABbiodiversity/wildRtrax-assets/main/utcoffset.tif", destinationPath = file.path(root, "Offset_prep"))
  
crs <- terra::crs(rtree)
```


#2. Data prep

Load data package from script 1 (i.e., data wrangled)
Note that this data does not have any NA on the latitude or longitude. 
```{r}
load(file.path(root, "1_RM_data_clean.R"))
```

Split into local and utc time zone objects

```{r}
visit.local <- visit %>% 
  dplyr::filter(source!="eBird")
visit.utc <- visit %>% 
  dplyr::filter(source=="eBird")
```


#3. Create functions

##3.1.Create prediction object

Function make_pred_obj to create a prediction object, where tz can be local (i.e., data coming from wildtrax) or utc (i.e., data coming from eBird). 
Data is going to be the visits created in step 1.3. 
```{r}

make_pred_obj <- function(data, tz="local") {
  
  #Extract varibles
  date <- word(data$date,1)
  time <- word(data$date,-1) 
  lon <- data$longitude
  lat <- data$latitude
  dur <- data$duration
  dis <- data$distance
  
  #parse date+time into POSIXlt
  if(tz=="local"){
    dtm <- strptime(paste0(date, " ", time, ":00"),
                    format="%Y-%m-%d %H:%M:%S", tz="America/Edmonton")
  }
  if(tz=="utc"){
    dtm <- strptime(paste0(date, " ", time, ":00"),
                    format="%Y-%m-%d %H:%M:%S", tz="GMT")
  }
  day <- as.integer(dtm$yday)
  hour <- as.numeric(round(dtm$hour + dtm$min/60, 2))
    
  
  #checkfun is a function to see if variables are properly within the range

  checkfun <- function(x, name="", range=c(-Inf, Inf)) {
    if (any(x[!is.na(x)] %)(% range))
      stop(sprintf("Parameter %s is out of range [%.0f, %.0f]", name, range[1], range[2]))
    invisible(NULL)
  }
  
  # Check that fields are within the range
  checkfun(day, "day", c(0, 365))
  checkfun(hour, "hour", c(0, 24))
  checkfun(dur, "dur", c(0, Inf))
  checkfun(dur, "dis", c(0, Inf))

  #Create a dataframe with the proper projection
  xydf <- data.frame(x=lon, y=lat)
  xy <- terra::vect(xydf, geom=c("x", "y"), crs="+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
  xy <- terra::project(xy, crs)

  #LCC4 and LCC2
  vlcc <- terra::extract(rlcc, xy)$lcc
  lcclevs <- c("0"="", "1"="Conif", "2"="Conif", "3"="", "4"="",
               "5"="DecidMixed", "6"="DecidMixed", "7"="", "8"="Open", "9"="",
               "10"="Open", "11"="Open", "12"="Open", "13"="Open", "14"="Wet",
               "15"="Open", "16"="Open", "17"="Open", "18"="", "19"="")
  lcc4 <- factor(lcclevs[vlcc+1], c("DecidMixed", "Conif", "Open", "Wet"))
  lcc2 <- lcc4
  levels(lcc2) <- c("Forest", "Forest", "OpenWet", "OpenWet")

  #TREE
  vtree <- terra::extract(rtree, xy)$tree
  TREE <- vtree / 100
  TREE[TREE %)(% c(0, 1)] <- 0

  #raster::extract seedgrow value (this is rounded)
  d1 <- terra::extract(rd1, xy)$seedgrow

  #UTC offset + 7 makes Alberta 0 (MDT offset) for local times
  if(tz=="local"){
    ltz <- terra::extract(rtz, xy)$utcoffset + 7
  }
  if(tz=="utc"){
    ltz <- 0
  }

  #sunrise time adjusted by offset
  ok_dt <- !is.na(dtm)
  dtm[is.na(dtm)] <- mean(dtm, na.rm=TRUE)
  if(tz=="local"){
    sr <- suntools::sunriset(cbind("X"=xydf$x, "Y"=xydf$y),
                   as.POSIXct(dtm, tz="America/Edmonton"),
                   direction="sunrise", POSIXct.out=FALSE) * 24
  }
  if(tz=="utc"){
    sr <- suntools::sunriset(cbind("X"=xydf$x, "Y"=xydf$y),
                   as.POSIXct(dtm, tz="GMT"),
                   direction="sunrise", POSIXct.out=FALSE) * 24
  }
  TSSR <- round(unname((hour - sr + ltz) / 24), 4)

  #days since local spring
  DSLS <- (day - d1) / 365

  #transform the other variables
  JDAY <- round(day / 365, 4) # 0-365
  TREE <- round(vtree / 100, 4)
  MAXDIS <- round(dis / 100, 4)
  MAXDUR <- round(dur, 4)

  pred_obj <- data.frame(
    TSSR=TSSR,
    JDAY=JDAY,
    DSLS=DSLS,
    LCC2=lcc2,
    LCC4=lcc4,
    TREE=TREE,
    MAXDUR=MAXDUR,
    MAXDIS=MAXDIS, 
    id=data$id)
  
  return(pred_obj)

}
```

##3.2.Create offset function

Offset can only be calculated for species included species consider by QPAD. 
spp can be an object including one or multiple species of interest. 
The full list can be call using getBAMspecieslist().
Note that x is this case the output generated in the function make_pred_obj.


```{r}
make_offset <- function(x, spp){

  if (length(spp) > 1L)
    stop("spp argument must be length 1. Use a loop or map for multiple species.")
  spp <- as.character(spp)

  #checks
  if (!(spp %in% getBAMspecieslist()))
    stop(sprintf("Species %s has no QPAD estimate available", spp))

  #constant for NA cases
  cf0 <- exp(unlist(coefBAMspecies(spp, 0, 0)))

  #best model
  mi <- bestmodelBAMspecies(spp, type="BIC")
  cfi <- coefBAMspecies(spp, mi$sra, mi$edr)

  TSSR <- x$TSSR
  DSLS <- x$DSLS
  JDAY <- x$JDAY
  lcc2 <- x$LCC2
  lcc4 <- x$LCC4
  TREE <- x$TREE
  MAXDUR <- x$MAXDUR
  MAXDIS <- x$MAXDIS
  n <- nrow(x)

  #Design matrices for singing rates (`Xp`) and for EDR (`Xq`)
  Xp <- cbind(
    "(Intercept)"=1,
    "TSSR"=TSSR,
    "JDAY"=JDAY,
    "TSSR2"=TSSR^2,
    "JDAY2"=JDAY^2,
    "DSLS"=DSLS,
    "DSLS2"=DSLS^2)

  Xq <- cbind("(Intercept)"=1,
              "TREE"=TREE,
              "LCC2OpenWet"=ifelse(lcc4 %in% c("Open", "Wet"), 1, 0),
              "LCC4Conif"=ifelse(lcc4=="Conif", 1, 0),
              "LCC4Open"=ifelse(lcc4=="Open", 1, 0),
              "LCC4Wet"=ifelse(lcc4=="Wet", 1, 0))

  p <- rep(NA, n)
  A <- q <- p

  #design matrices matching the coefs
  Xp2 <- Xp[,names(cfi$sra),drop=FALSE]
  OKp <- rowSums(is.na(Xp2)) == 0
  Xq2 <- Xq[,names(cfi$edr),drop=FALSE]
  OKq <- rowSums(is.na(Xq2)) == 0

  #calculate p, q, and A based on constant phi and tau for the respective NAs
  p[!OKp] <- sra_fun(MAXDUR[!OKp], cf0[1])
  unlim <- ifelse(MAXDIS[!OKq] == Inf, TRUE, FALSE)
  A[!OKq] <- ifelse(unlim, pi * cf0[2]^2, pi * MAXDIS[!OKq]^2)
  q[!OKq] <- ifelse(unlim, 1, edr_fun(MAXDIS[!OKq], cf0[2]))

  #calculate time/lcc varying phi and tau for non-NA cases
  phi1 <- exp(drop(Xp2[OKp,,drop=FALSE] %*% cfi$sra))
  tau1 <- exp(drop(Xq2[OKq,,drop=FALSE] %*% cfi$edr))
  p[OKp] <- sra_fun(MAXDUR[OKp], phi1)
  unlim <- ifelse(MAXDIS[OKq] == Inf, TRUE, FALSE)
  A[OKq] <- ifelse(unlim, pi * tau1^2, pi * MAXDIS[OKq]^2)
  q[OKq] <- ifelse(unlim, 1, edr_fun(MAXDIS[OKq], tau1))

  #log(0) is not a good thing, apply constant instead
  ii <- which(p == 0)
  p[ii] <- sra_fun(MAXDUR[ii], cf0[1])

  #package output
  data.frame(
    p=p,
    q=q,
    A=A,
    correction=p*A*q,
    offset=log(p) + log(A) + log(q))

}
```

##3.2.Create qpad offset function

to put together the two previous functions. 
Note that the object x is the pred_obj.
```{r}
qpad_offsets <- function(x, species = c("all")) {

  #Make the species list
  if("all" %in% species) {
    spp <- sort(getBAMspecieslist()) 
  }
  
  else {
    spp <- species 
  }
   
  #Set up the offset loop
  off <- matrix(0, nrow(x), length(spp))
  colnames(off) <- spp

  #Make the offsets
  for (i in 1:length(spp)){
    cat("\n", spp[i])
    o <- make_offset(x, spp[i])
    off[,i] <- o$offset
  }

  #Put together
  
  out <- cbind(x %>%
                 select (id), 
               data.frame(off) %>%
                 rename_with(.fn=~paste0(.x, ".off"))) 
    
    
  
  return(out)
}

```

#4. Run the function 
```{r}

#Make prediction object
  
data.local <- make_pred_obj(visit.local, tz="local")
data.utc <- make_pred_obj(visit.utc, tz="utc")

#Make offsets

offset.local<-qpad_offsets(data.local)
offset.utc<-qpad_offsets(data.utc)

#Put together local and utc

offsets<-rbind(offset.local,offset.utc)
```

#5. Save
```{r}
save(offsets, file=file.path(root, "2_RM_data_offsets.R"))
```
